% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_parser.R
\name{create_parser}
\alias{create_parser}
\title{create_parser}
\usage{
create_parser()
}
\value{
An instance of type 'optparse::OptionParser'.
}
\description{
Creates an object that defines and handles command line arguments.
}
\details{
A parser that organizes the communication between the user and th function.
It also provides a help message.
}
\author{
Sebastian Malkusch
}
