% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_resample.R
\name{create_resample_experiment}
\alias{create_resample_experiment}
\title{create_resample_experiment}
\usage{
create_resample_experiment(
  seed,
  data_df,
  parser_inst,
  model_inst,
  config_inst,
  n_features
)
}
\arguments{
\item{seed}{sets the seed for the random number generator to guarantee reproducibility.
(int)}

\item{data_df}{data frame to be learned from.
(tibble::tibble)}

\item{parser_inst}{instance of parser object. (optparse::parse_args).}

\item{model_inst}{instance of caret_train object (caret::train).}

\item{config_inst}{list of config options (list).}

\item{n_features}{number of features (int).}
}
\value{
An instance of type 'Resampler'.
}
\description{
Creates an object of a resampling experiment.
}
\details{
Creates a resampling experiment.
It uses user defined parameters to set up the experiment.
It creates an instance of the Resampler object
and runs the experiment according to the user-defined parameters.
}
\author{
Sebastian Malkusch
}
