% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_bootstrap.R
\name{fml_bootstrap}
\alias{fml_bootstrap}
\title{fml_bootstrap}
\usage{
fml_bootstrap(parser_inst)
}
\arguments{
\item{parser_inst}{Instance of fml_parser class that comprises command line arguments.}
}
\value{
none
}
\description{
Pipeline function that sets up and runs a resampling experiment.
}
\details{
The experiment is run in parallel.
All results are written to files.
}
\examples{
\dontrun{
  flowml::fml_bootstrap(parser_inst)
}

}
\author{
Sebastian Malkusch
}
