% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_example.R
\name{fml_example}
\alias{fml_example}
\title{fml_example}
\usage{
fml_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
The path of to an example file, if file is defined.
Else, a list of example files.
}
\description{
path to flowml examples data
}
\details{
flowml comes bundled with a number of sample files in its
`inst/extdata` directory. This function allows to access them.
}
\examples{
\dontrun{
fml_example()
fml_example(file = "reg_config.json")
}

}
\author{
Sebastian Malkusch
}
