% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_interpret.R
\name{fml_interpret}
\alias{fml_interpret}
\title{fml_interpret}
\usage{
fml_interpret()
}
\description{
Pipeline function that sets up and runs a post-hoc interpretation of an ml experiment.
All results are written to rds files.
}
\author{
Sebastian Malkusch
}
