% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_train.R
\name{fml_train}
\alias{fml_train}
\title{fml_train}
\usage{
fml_train(parser_inst)
}
\arguments{
\item{parser_inst}{instance of fml_parser class that comprises command line arguments.}
}
\description{
Pipeline function that performs a hyperparameter screeing experiment.
}
\examples{
\dontrun{
  flowml::fml_train(parser_inst)
}

}
\author{
Kolja Becker
}
