% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_validate.R
\name{fml_validate}
\alias{fml_validate}
\title{fml_validate}
\usage{
fml_validate(parser_inst)
}
\arguments{
\item{parser_inst}{instance of fml_parser class that comprises command line arguments.}
}
\value{
none
}
\description{
Pipeline function that performs a validation experiment on a
a caret train object based on test samples.
}
\examples{
\dontrun{
  flowml::fml_validate(parser_inst)
}

}
\author{
Kolja Becker
}
