% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_format_response.R
\name{format_y}
\alias{format_y}
\title{format_y}
\usage{
format_y(y, ml.type)
}
\arguments{
\item{y}{vector of response varibale.}

\item{ml.type}{type of experiment (chracater).}
}
\value{
a transformed version of the response variable y.
}
\description{
Formats response variable based on the ml-type variable passed by the config file.
For regression analyses the response variable will be explicitly transformed to type numeric.
For Classification experiments the response variable will be explicitly transformed to a factor.
Time-to-event models are to be implemented in the near future.
}
\author{
Kolja Becker
}
