% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_categorize.R
\name{run_abc_analysis}
\alias{run_abc_analysis}
\title{Performs item categorization}
\usage{
run_abc_analysis(data_obj, method)
}
\arguments{
\item{data_obj}{Resuls of model interpretation experiment}

\item{method}{Method used for model interpretation (permutatopn or shap)}
}
\value{
A tibble with item categories

a tibble
}
\description{
Performs item categorization on permutation or shap analysis object
}
\details{
Interpretation results are passed to the function.
Based on the type of interpretation experiment the data is transformed into a uniformly structured data frame.
Item categorization is performed by computed ABC analysis.
The result is returned in form of a tibble.
}
\author{
Sebastian Malkusch
}
