% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_bootstrap.R
\name{fml_bootstrap}
\alias{fml_bootstrap}
\title{fml_bootstrap}
\usage{
fml_bootstrap(parser_inst)
}
\arguments{
\item{parser_inst}{Instance of fml_parser class that comprises command line arguments.}
}
\value{
none
}
\description{
Pipeline function that sets up and runs a resampling experiment.
}
\details{
The experiment is run in parallel.
All results are written to files.
}
\examples{
\dontrun{
  Rscript fml_bootstrap.R --config /path/to/config.json --data /path/to/data.csv --samples /path/to/samples_train.txt --features path/to/features.txt --trained /path/to/trained_model.rds --permutation none --cores 1
  Rscript fml_bootstrap.R --config /path/to/config.json --data /path/to/data.csv --samples /path/to/samples_train.txt --features path/to/features.txt --trained /path/to/trained_model.rds --permutation response --cores 1
  Rscript fml_bootstrap.R --config /path/to/config.json --data /path/to/data.csv --samples /path/to/samples_train.txt --features path/to/features.txt --trained /path/to/trained_model.rds --permutation features --cores 1 --extended_features /path/to/extended_features.txt
}

}
\author{
Sebastian Malkusch
}
\keyword{bootstrap,}
\keyword{hypothesis}
\keyword{permutation,}
\keyword{resmapling,}
\keyword{testing}
