% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml_interpret.R
\name{fml_interpret}
\alias{fml_interpret}
\title{fml_interpret}
\usage{
fml_interpret(parser_inst)
}
\arguments{
\item{parser_inst}{instance of fml_parser class that comprises command line arguments.}
}
\description{
Pipeline function that sets up and runs a post-hoc interpretation of an ml experiment.
All results are written to rds files.
}
\examples{
\dontrun{
  flowml::fml_interpret(parser_inst)
}

}
\author{
Sebastian Malkusch
}
